+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Enhancement: Add SentryWidgetsBindingObserver, an Integration that captures certain window and device events.
 - Ref: Only assign non-null option values in Android native integration in order preserve default values
 - Enhancement: Add 'attachThreads' in options. When enabled, threads are attached to all logged events for Android
+- Ref: Rename typedef `Logger` to `SentryLogger` to prevent name clashes with logging packages
 
 ## 4.0.0-beta.1